+++ tests/moneybin/matching/test_transfer_integration.py
@@ -17,70 +17,8 @@ from moneybin.matching.persistence import (
 from moneybin.matching.reconciliation import retire_transfers_invalidated_by_dedup
 from moneybin.repositories.match_decisions_repo import MatchDecisionsRepo
 from moneybin.services.matching_service import MatchingService
-
-
-def _setup_tables(db: Database) -> None:
-    """Create the unioned table and match_decisions for integration tests."""
-    db.execute("CREATE SCHEMA IF NOT EXISTS app")
-    db.execute("""
-        CREATE TABLE IF NOT EXISTS app.match_decisions (
-            match_id VARCHAR NOT NULL,
-            source_transaction_id_a VARCHAR NOT NULL,
-            source_type_a VARCHAR NOT NULL,
-            source_origin_a VARCHAR NOT NULL,
-            source_transaction_id_b VARCHAR NOT NULL,
-            source_type_b VARCHAR NOT NULL,
-            source_origin_b VARCHAR NOT NULL,
-            account_id VARCHAR NOT NULL,
-            confidence_score DECIMAL(5, 4),
-            match_signals JSON,
-            match_type VARCHAR NOT NULL DEFAULT 'dedup',
-            match_tier VARCHAR,
-            account_id_b VARCHAR,
-            match_status VARCHAR NOT NULL,
-            match_reason VARCHAR,
-            decided_by VARCHAR NOT NULL,
-            decided_at TIMESTAMP NOT NULL,
-            reversed_at TIMESTAMP,
-            reversed_by VARCHAR,
-            PRIMARY KEY (match_id)
-        )
-    """)
-    db.execute("""
-        CREATE OR REPLACE TABLE _test_unioned (
-            source_transaction_id VARCHAR,
-            account_id VARCHAR,
-            transaction_date DATE,
-            amount DECIMAL(18, 2),
-            description VARCHAR,
-            source_type VARCHAR,
-            source_origin VARCHAR,
-            source_file VARCHAR,
-            currency_code VARCHAR DEFAULT 'USD'
-        )
-    """)
-
-
-def _insert(
-    db: Database,
-    stid: str,
-    acct: str,
-    txn_date: str,
-    amount: str,
-    desc: str,
-    stype: str = "csv",
-    sorigin: str = "bank",
-    sfile: str = "test.csv",
-) -> None:
-    db.execute(
-        """
-        INSERT INTO _test_unioned (
-            source_transaction_id, account_id, transaction_date, amount,
-            description, source_type, source_origin, source_file
-        ) VALUES (?, ?, ?::DATE, ?::DECIMAL(18,2), ?, ?, ?, ?)
-        """,
-        [stid, acct, txn_date, amount, desc, stype, sorigin, sfile],
-    )
+from tests.moneybin.matching.conftest import create_matcher_tables as _setup_tables
+from tests.moneybin.matching.conftest import insert_matcher_row as _insert
 
 
 @pytest.mark.integration
